Frontend/backend/requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
audioop-lts==0.2.2
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.14.3
bitsandbytes==0.49.1
black==26.1.0
blinker==1.9.0
boto3==1.42.41
botocore==1.42.41
brotli==1.2.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.0
click==8.3.1
colorama==0.4.6
comtypes==1.4.12
cryptography==46.0.4
dataclasses-json==0.6.7
datasets==4.0.0
ddgs==9.5.5
deepface==0.0.96
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docopt==0.6.2
docutils==0.22.4
duckduckgo_search==8.1.1
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
faiss-cpu==1.12.0
fastapi==0.110.1
fastuuid==0.14.0
ffmpy==1.0.0
filelock==3.20.3
fire==0.7.1
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2026.1.0
gast==0.7.0
gdown==5.2.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.61.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
gradio==6.1.0
gradio_client==2.0.1
greenlet==3.2.4
groovy==0.1.2
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface_hub==1.3.7
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
jaraco.context==6.1.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.2
jq==1.11.0
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kernels==0.12.1
keyboard==0.13.5
kubernetes==35.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-text-splitters==0.3.11
langsmith==0.4.21
libclang==18.1.1
librt==0.7.8
lightecc==0.0.4
lightphe==0.0.20
litellm==1.80.0
llama_cpp_python==0.3.16
lxml==6.0.1
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
more-itertools==10.8.0
motor==3.3.1
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.7.1
multiprocess==0.70.16
mypy==1.19.1
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.5
nh3==0.3.2
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.1
openai==1.99.9
opencv-python==4.12.0.88
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.3
overrides==7.7.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
pefile==2024.8.26
pillow==12.1.0
pipwin==0.5.2
platformdirs==4.5.1
pluggy==1.6.0
posthog==5.4.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.5
psutil==7.0.0
pvporcupine==4.0.0
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
pybase64==1.4.3
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.10.1
pydantic_core==2.41.5
pydub==0.25.1
pyflakes==3.4.0
pygame==2.6.1
Pygments==2.19.2
pyjsparser==2.7.1
PyJWT==2.11.0
pymongo==4.5.0
pyparsing==3.3.2
PyPika==0.51.1
pypiwin32==223
PyPrind==2.11.3
pyproject_hooks==1.2.0
pySmartDL==1.3.4
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.1
pyttsx3==2.99
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retina-face==0.0.17
rfc3986==2.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
safehttpx==0.1.7
safetensors==0.6.2
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
spark-parser==1.8.9
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.37.2
stripe==14.3.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomlkit==0.13.3
tqdm==4.67.3
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uncompyle6==3.9.2
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.25.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
wikipedia==1.4.0
wrapt==2.0.1
xdis==6.1.5
xxhash==3.5.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.24.0
chromadb
